backend/requirements.txt
alembic==1.4.3
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-JWT-Extended==3.24.1
flask-redis==0.4.0
flask-restful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
marshmallow-sqlalchemy==0.23.1
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
PyYaml==5.3.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.19
Werkzeug==1.0.1
